extract data validation into mixin classes
